PIPELINE: fix issues from codeFactor Replaced original dicom2glb (GDCM) pipeline with sitk version completely
