+++ HoloPipelines/pipelineList.json
@@ -19,18 +19,9 @@
       },
    "dicom2glb":
       {
-         "name":"DICOM to glb generation (Pydicom)",
+         "name":"DICOM to glb generation",
          "src":"pipelines/pipelineDicom2glb.py",
          "info":"Generate glb mesh from dicom",
-         "addDate":"01/07/2019",
-         "modDate":"01/07/2019",
-         "param":"3"
-      },
-   "dicom2glb-sitk":
-      {
-         "name":"DICOM to glb generation (Simple ITK)",
-         "src":"pipelines/pipelineDicom2glb_sitk.py",
-         "info":"Generate glb mesh from dicom. Note: this pipeline uses Simple ITK to load Dicom pixel data instead of Pydicom",
          "addDate":"01/07/2019",
          "modDate":"01/07/2019",
          "param":"3"
